# docs/swagger.yaml
basePath: /
definitions:
...
  models.JSONError:
    properties:
      code:
        description: HTTP код ошибки
        example: 400
        type: integer
      message:
        description: Сообщение об ошибке
        example: Error message
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для безопасного хранения секретов и файлов
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GophKeeper API
  version: "1.0"
paths:
...
  /api/files/{label}/download:
    get:
      consumes:
      - application/json
      description: Скачивает зашифрованный файл с сервера по метке
      parameters:
      - description: Метка файла
        in: path
        name: label
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Файл успешно скачан
          schema:
            type: file
        "400":
          description: Некорректные данные запроса
          schema:
            $ref: '#/definitions/models.JSONError'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.JSONError'
        "404":
          description: Файл не найден
          schema:
            $ref: '#/definitions/models.JSONError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.JSONError'
      security:
      - ApiKeyAuth: []
      summary: Скачивание файла
      tags:
      - files
...
securityDefinitions:
  ApiKeyAuth:
    description: Токен доступа в формате Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...